# user-service/src/main/java/com/ecommerce/user/infrastructure/adapter/in/web/dto/LoginRequest.java
package com.ecommerce.user.infrastructure.adapter.in.web.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(

        @NotBlank(message = "L'email est obligatoire")
        @Email(message = "Format email invalide")
        String email,

        @NotBlank(message = "Le mot de passe est obligatoire")
        String password
) {}