# user-service/src/main/resources/openapi/user-service-api.yml
openapi: 3.0.3
info:
  title: User Service API
  description: |
    API du microservice User — Application E-Commerce Modulaire.
    Gère l'authentification, la création de comptes et la gestion des profils.
  version: 1.0.0
  contact:
    name: Mohammed Riad Moutaoukil

servers:
  - url: http://localhost:8081
    description: Serveur de développement local

tags:
  - name: Auth
    description: Authentification et génération de JWT Token
  - name: Users
    description: Gestion des utilisateurs et profils

paths:

  # ===== AUTH =====

  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Inscription d'un nouvel utilisateur
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              firstName: Mohammed Riad
              lastName: Moutaoukil
              email: riad@example.com
              password: Password123!
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Connexion d'un utilisateur existant
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: riad@example.com
              password: Password123!
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===== USERS =====

  /api/users/me:
    get:
      tags:
        - Users
      summary: Récupérer le profil de l'utilisateur connecté
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Token JWT manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Mettre à jour le profil de l'utilisateur connecté
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profil mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token JWT manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Récupérer un utilisateur par son ID (admin uniquement)
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant unique de l'utilisateur
      responses:
        '200':
          description: Utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Token JWT manquant ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Accès refusé — rôle ADMIN requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ===== COMPONENTS =====

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Token obtenu via POST /api/auth/login"

  schemas:

    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          example: Mohammed Riad
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          example: Moutaoukil
        email:
          type: string
          format: email
          example: riad@example.com
        password:
          type: string
          minLength: 8
          maxLength: 100
          description: "Doit contenir au moins 1 majuscule, 1 chiffre et 1 caractère spécial"
          example: Password123!

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: riad@example.com
        password:
          type: string
          example: Password123!

    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          example: Mohammed Riad
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          example: Moutaoukil
        phoneNumber:
          type: string
          example: "+33612345678"

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Token à utiliser dans les requêtes suivantes
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Durée de validité en secondes
          example: 86400
        user:
          $ref: '#/components/schemas/UserResponse'

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        firstName:
          type: string
          example: Mohammed Riad
        lastName:
          type: string
          example: Moutaoukil
        email:
          type: string
          format: email
          example: riad@example.com
        phoneNumber:
          type: string
          example: "+33612345678"
        role:
          type: string
          enum:
            - USER
            - ADMIN
          example: USER
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Utilisateur introuvable
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        path:
          type: string
          example: /api/users/123